restructure repo and inplace hovercrestion(to be refactored back ;))
