+++ src/assetmanagement.rs
@@ -4,9 +4,10 @@ use amethyst::{
 };
 use std::collections::HashMap;
 
-/// This trait provides basic information about the assets on disk
+/// This (private!) trait provides basic information about the assets on disk
 /// The data is stored in a dedicated folder per asset (e.g. `.png`, `.obj`, ...). Each of which has to be identical to the other asset folders except for the different endings
-pub trait AssetInformation {
+#[allow(private_in_public)] // This is a trait, that should not be seen outside of this file.
+trait AssetInformation {
     fn folder_name(&self) -> &'static str;
     fn extension(&self) -> &'static str;
 }
@@ -96,5 +97,75 @@ where
         );
         self.assets.insert(String::from(path), handle.clone());
         handle
+    }
+}
+
+pub mod util {
+    use amethyst::{
+        assets::{AssetStorage, Loader},
+        ecs::{Entity, ReadExpect, Write, WriteStorage},
+        renderer::{
+            Material, MaterialDefaults, Mesh, MeshHandle, ObjFormat, PngFormat, Texture,
+            TextureMetadata,
+        },
+    };
+    use assetmanagement::AssetManager;
+
+    /// A collection of the required resources necessary to attach `Texture` and `MeshHandle` to an entity
+    pub type AssetStorages<'a> = (
+        ReadExpect<'a, Loader>,
+        Write<'a, AssetManager<Mesh>>,
+        WriteStorage<'a, MeshHandle>,
+        Write<'a, AssetStorage<Mesh>>,
+        Write<'a, AssetManager<Texture>>,
+        WriteStorage<'a, Material>,
+        Write<'a, AssetStorage<Texture>>,
+        ReadExpect<'a, MaterialDefaults>,
+    );
+
+    /// Attaches `Material` and `MeshHandle` of specified asset to the entity
+    pub fn insert_into_asset_storages(
+        entity: Entity,
+        asset_name: &str,
+        storages: &mut AssetStorages,
+    ) {
+        let (
+            ref loader,
+            ref mut mesh_manager,
+            ref mut mesh_handles,
+            ref mut mesh_storage,
+            ref mut tex_manager,
+            ref mut mat_storage,
+            ref mut tex_storage,
+            ref default_mat,
+        ) = storages;
+
+        let mesh = {
+            mesh_manager.get_asset_handle_or_load(
+                asset_name,
+                ObjFormat,
+                Default::default(),
+                mesh_storage,
+                &loader,
+            )
+        };
+
+        // load texture/material
+        let material = {
+            let handle = tex_manager.get_asset_handle_or_load(
+                asset_name,
+                PngFormat,
+                TextureMetadata::srgb(),
+                tex_storage,
+                &loader,
+            );
+            Material {
+                albedo: handle,
+                ..default_mat.0.clone()
+            }
+        };
+
+        mat_storage.insert(entity, material).unwrap();
+        mesh_handles.insert(entity, mesh).unwrap();
     }
 }
