+++ src/eventhandling/mod.rs
@@ -1,3 +1,9 @@
-mod click_handling;
+mod camera_movement;
+mod clicking;
+mod hovering;
+mod mouse_ray;
 
-pub use self::click_handling::*;
+pub use self::camera_movement::CameraMovementSystem;
+pub use self::clicking::*;
+pub use self::hovering::*;
+pub use self::mouse_ray::*;
