+++ src/systems/mod.rs
@@ -1,13 +1,7 @@
-mod camera_movement;
 mod ground_reveal;
-mod hover_interaction;
-mod mouse_ray;
 mod movement;
 mod oxygen;
 
-pub use self::camera_movement::CameraMovementSystem;
 pub use self::ground_reveal::{GroundRevealSystem, RevealQueue};
-pub use self::hover_interaction::*;
-pub use self::mouse_ray::*;
 pub use self::movement::{MovementSystem, Path};
 pub use self::oxygen::{Oxygen, OxygenBar, OxygenSystem};
